requirements.txt
crewai
gradio
requests
python-dotenv
openai
pandas
PyPDF2
pytest